requirements.txt
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
Flask==2.0.1
Flask-PyMongo==2.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pip==21.1.2
pymongo==3.11.4
python-dotenv==0.17.1
setuptools==57.0.0
Werkzeug==2.0.1
gunicorn=20.1.0
